+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 ### Added
+- Send entity operations events to all registered plugins
 ### Changed
 ### Fixed
 ### Removed